requirements.txt
filelock==3.12.0
fsspec==2023.5.0
ftfy==6.1.1
huggingface-hub==0.21.2
open-clip-torch==2.10.1
packaging==23.1
protobuf==3.20.3
pyyaml==6.0
regex==2023.6.3
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.99
timm==0.9.2
wcwidth==0.2.6
datasets==2.19.0
wandb==0.16.6
hydra-core==1.3.2
h5py==3.11.0
gdown==5.1.0
